Update TriangleMarker.js
Same as the former "CurrentCoordinate.js": 
Old version of "TriangleMarker.js" is not working for svg-mode, if "stroke" or "fill" is a function. This version works.

I also added property "direction". With this, the triangle-top can be positioned (top, bottom) or hide for some points. It also accepts a function. Actually, in this version, the "direction"-property only works for svg-mode!! In  canvas-mode, I could not rotate the triangle. Please look in "function Triangle.drawOnCanvasWithNoStateChange" - there is a "TODO"! Maybe, you can correct the code to make the triangle rotation also works for canvas-mode.