Update CrossHairCursor.js
- New property: "disableLineY", so y-crosshair can be disabled.

- New property: "expandStrokeWidthLineX" TODO (this property does not work). Please help to implement this feature:

TODO: if "expandStrokeWidthLineX" is true, 
then "stroke-width" of the x-axis-crosshair should expand to its XAxis-tickSize (?).
For example, on 15 minute chart, the stroke-width of the x-axis-crosshair should be something like "15".
So the width of the x-axis-crosshair should be the same as the width of the candle. How can I do that?